Make file input not composable
